refactor: patterns list
