+++ src/app.css
@@ -116,7 +116,7 @@ div.wpui_column {
 }
 
 .copied_pop .components-popover__content {
-    padding: 8px !important
+    padding: 6px !important
 }
 
 .wpui_view_toggle [role="presentation"] {
