+++ src/components/copy-button.js
@@ -1,0 +1,51 @@
+/**
+ * External dependencies.
+ */
+import { useState } from 'react';
+
+/**
+ * WordPress dependencies.
+ */
+import { useCopyToClipboard } from '@wordpress/compose';
+import {
+    check,
+    copy,
+} from "@wordpress/icons";
+import {
+    Popover,
+    Button,
+} from "@wordpress/components";
+
+/**
+ * Render Copy Button
+ */
+function CopyButton({ content }) {
+    const [hasCopied, setHasCopied] = useState(false);
+
+    const handleCopyClick = () => {
+        setHasCopied(true);
+        setTimeout(() => {
+            setHasCopied(false);
+        }, 1000);
+    };
+
+    return (
+        <Button
+            className='wpui_copy'
+            icon={hasCopied ? check : copy}
+            onClick={handleCopyClick}
+            ref={useCopyToClipboard(content)}
+        >
+            {hasCopied && (
+                <Popover
+                    className='copied_pop'
+                    position='middle left'
+                >
+                    Copied
+                </Popover>
+            )}
+        </Button>
+    );
+}
+
+export default CopyButton;
