+++ src/components/index.js
@@ -1,11 +1,13 @@
 /**
- * Cards
+ * Components
  */
 export { default as Header } from './header';
 export { default as Footer } from './footer';
 export { default as Banner } from './banner';
 export { default as Logo } from './logo';
+export { default as CopyButton } from './copy-button';
 export { default as ContentLoader } from './content-loader';
 export { default as ScrollToTop } from './scroll-to-top';
 export { default as CategoryCard } from './category-card';
-export { default as PatternList } from './pattern-list';
+export { default as PatternView } from './pattern-view';
+export { default as PatternCode } from './pattern-code';
