--- src/data/index.js
@@ -1,5 +1,0 @@
-/**
- * Export Data
- */
-export { default as WpuiProvider } from './context';
-export { WpuiContext } from './context';
